Add files via upload
This is really the backbone of what I want to do with the bot. All necessarry information for cards is stored in player_cards.py. The actual initialization of the bot and message detection is in HOPS.py. commands.py make up a majority of the project, and all of the really important commands have already been implemented as functions. First, all the logic for the database. Second, everything needed to send a set of cards to a user, three random cards are picked, their images are compiled, then they're sent, giving the player a set amount of time to pick using a reaction. Third, players need to be able to view their collection. Displaying the player's cards on one page, and using reactions to switch between them. Collection traversal isn't perfected yet, but most other problems have been ironed out. This is a good place to start before I start implementing the more interactive elements the bot.